+++ RELEASE_NOTES.md
@@ -1,5 +1,23 @@
 # Release Notes
 
+## Version TBD
+_Date TBD_
+
+### Table modifications
+* Add reason column to some questionable activity tables that can result from listing update
+* Add fuzzy_choices table
+* Add FK from pending_certified_product to certified_product table to track listings from pending to confirmed
+
+---
+
+## Version 13.2.1
+_29 January 2018_
+
+### Data modifications
+* Insert some CMS IDs that were lost due to database issues
+
+---
+
 ## Version 13.2.0
 _17 January 2018_